# Fuse q, k, v matmuls