@karpathy 2023-07-25
This is quite nice, sadly it bloats the code quite a bit in terms of LOC.... hard