@karpathy 2023-07-25
wow, very interesting. is this somehow tiling the qkv matmul into 4 * 4 = 16 nested loops? so it's 16X unrolled?