@Maknee 2023-07-25
Reduced the LOC. Performance uplift varies from original.

Old:

| | % uplift (model) | % uplift (model44m) |
| - |----------- | ----------- |
| run | 1.08 | 1.13|
| runfast| 1.06| 1.05|
| runomp| 1.07 | 1.06|


New:

| | % uplift (model) | % uplift (model44m) |
| - |----------- | ----------- |
| run | 1.08 | 1.14|
| runfast| 1.04| 1.01|
| runomp| 1.02| 1.13|

---

## run

```
Benchmark 1: ./run model.bin 0.0
  Time (mean ± σ):      2.030 s ±  0.002 s    [User: 2.010 s, System: 0.020 s]
  Range (min … max):    2.028 s …  2.032 s    10 runs
 
Benchmark 2: ./run_qkv model.bin 0.0
  Time (mean ± σ):      1.875 s ±  0.002 s    [User: 1.855 s, System: 0.020 s]
  Range (min … max):    1.873 s …  1.877 s    10 runs
 
Summary
  ./run_qkv model.bin 0.0 ran
    1.08 ± 0.00 times faster than ./run model.bin 0.0
```

---

```
Benchmark 1: ./run model44m.bin 0.0
  Time (mean ± σ):      6.301 s ±  0.128 s    [User: 6.243 s, System: 0.057 s]
  Range (min … max):    6.151 s …  6.402 s    10 runs

Benchmark 2: ./run_qkv model44m.bin 0.0
  Time (mean ± σ):      5.527 s ±  0.056 s    [User: 5.470 s, System: 0.057 s]
  Range (min … max):    5.483 s …  5.624 s    10 runs
 
Summary
  ./run_qkv model44m.bin 0.0 ran
    1.14 ± 0.03 times faster than ./run model44m.bin 0.0
```

## runfast

```
Benchmark 1: ./run model.bin 0.0
  Time (mean ± σ):     496.7 ms ±   5.5 ms    [User: 478.9 ms, System: 17.7 ms]
  Range (min … max):   488.8 ms … 504.0 ms    10 runs
 
Benchmark 2: ./run_qkv model.bin 0.0
  Time (mean ± σ):     478.6 ms ±   2.8 ms    [User: 457.8 ms, System: 20.8 ms]
  Range (min … max):   475.5 ms … 484.5 ms    10 runs
 
Summary
  ./run_qkv model.bin 0.0 ran
    1.04 ± 0.01 times faster than ./run model.bin 0.0
```

---

```
Benchmark 1: ./run model44m.bin 0.0
  Time (mean ± σ):      1.742 s ±  0.001 s    [User: 1.684 s, System: 0.058 s]
  Range (min … max):    1.741 s …  1.743 s    10 runs
 
Benchmark 2: ./run_qkv model44m.bin 0.0
  Time (mean ± σ):      1.729 s ±  0.004 s    [User: 1.670 s, System: 0.058 s]
  Range (min … max):    1.723 s …  1.736 s    10 runs
 
Summary
  ./run_qkv model44m.bin 0.0 ran
    1.01 ± 0.00 times faster than ./run model44m.bin 0.0
```

## runomp

```
Benchmark 1: ./run model.bin 0.0
  Time (mean ± σ):      61.9 ms ±   0.9 ms    [User: 367.3 ms, System: 21.0 ms]
  Range (min … max):    60.9 ms …  63.1 ms    10 runs
 
Benchmark 2: ./run_qkv model.bin 0.0
  Time (mean ± σ):      62.9 ms ±   0.8 ms    [User: 378.3 ms, System: 16.0 ms]
  Range (min … max):    61.7 ms …  64.3 ms    10 runs
 
Summary
  ./run model.bin 0.0 ran
    1.02 ± 0.02 times faster than ./run_qkv model.bin 0.0
```

---

```
Benchmark 1: ./run model44m.bin 0.0
  Time (mean ± σ):     585.8 ms ±   5.6 ms    [User: 4262.3 ms, System: 60.5 ms]
  Range (min … max):   580.2 ms … 596.8 ms    10 runs
 
Benchmark 2: ./run_qkv model44m.bin 0.0
  Time (mean ± σ):     517.9 ms ±   1.6 ms    [User: 3716.6 ms, System: 62.5 ms]
  Range (min … max):   515.8 ms … 519.6 ms    10 runs
 
Summary
  ./run_qkv model44m.bin 0.0 ran
    1.13 ± 0.01 times faster than ./run model44m.bin 0.0
```