@Maknee 2023-07-25
> wow, very interesting. is this somehow tiling the qkv matmul into 4 * 4 = 16 nested loops? so it's 16X unrolled?

Seems so, looking at the [generated assembly in godbolt](https://godbolt.org/z/f4vGe1ac7)