@krzysztof-jusiak 2023-07-25
Very clever change. It doesn't add too much complexity but I don't think it will be understood by everyone, if that's the goal of the project (just pointing it out, not saying that it shouldn't get in if it's worth it). If I'm reading it correctly though it seems that the performance gain is small looking at the average runs, is the tok/s much better? On the other hand, I also think it's a great example of why declarative computation graphs are so powerful, having easy to follow user interface and fused/optimized back-end is extremely powerful concept but that' defo not 500 lines of easy to follow code.