@axrwl 2023-07-25
Clever indeed. #75 should also show similar improvements in performance.