@karpathy 2023-07-27
Thank you this is clever but I think for now I want to try to keep a single matmul function, which can become a single point of heavy optimization.